Log Python warnings correctly
So there was this annoying "extern.louie will soon be deprecated in
favor of coherence.dispatcher" warning at every start. At first the
solution seemed easy:

with warnings.catch_warnings():
    warnings.simplefilter("ignore")
    c = Coherence()
    cp = ControlPoint(c)
    ...

But, louie (that's Coherence's signal dispatcher, by the way) doesn't
warn when it is loaded or so but on the first call of *a function* of
that module which happens deep in Coherence's core in a callback called
by Twisted. So no luck with that.

So I came across logging.captureWarnings() which looked reasonable, now
that we use Python's logging. If captureWarnings() is in effect though,
all warnings get logged by the 'py.warnings' logger rather than by
'coherence.extern.louie' or so, and defaulting -L py.warnings=critical
or so comes up to disabling warnings globally, which we do not want.

But there's record.module! Yes, if it just contained the *whole* module
name rather than just 'louie'. And it's a str, no module object. So no
luck with that either. I ended up guessing the module name by trying to
get the relative path inside the longest sys.path component and
reformatting it. That gives us 'coherence.extern.louie' for which we can
call getLogger().isEnabledFor() and if it's enabled replace
'py.warnings' by 'coherence.extern.louie' in the actual output.

And if not? Uh. We have no control in the Formatter over how the record
is being processed at all or by whom. Trying the obvious 'return None'
lead to 'None' on stderr, so no luck. We now have this
SkippableStreamHandler which calls self.format (which ends up in
StreamHandler and eventually in LogFormatter.format), checks if the
return value is None and only if not calls StreamHandler.emit (which
then calls LogFormatter.format again, therefore the caching).